requirements.txt
openai
python-dotenv
PyYAML
pytest
typer